# Performance: Complete Bookings Pagination Modernization