--- a/frontend/templates/bookings/list.html
+++ b/frontend/templates/bookings/list.html
@@ -113,6 +113,23 @@
     </div>
 </div>
 
+<!-- Pagination -->
+<div id="pagination" class="d-flex justify-content-between align-items-center mt-3">
+    <div class="pagination-info">
+        Показано <span id="pageStart">0</span>-<span id="pageEnd">0</span> из <span id="totalItems">0</span>
+    </div>
+    <nav>
+        <ul class="pagination">
+            <li class="page-item disabled">
+                <a class="page-link" href="#" id="prevPage">Назад</a>
+            </li>
+            <li class="page-item disabled">
+                <a class="page-link" href="#" id="nextPage">Вперед</a>
+            </li>
+        </ul>
+    </nav>
+</div>
+
 <!-- New Booking Modal -->
 <div class="modal fade" id="newBookingModal" tabindex="-1">
     <div class="modal-dialog modal-lg">